requirements_old.txt
pandas
python-rapidjson
numba==0.55.1
requests==2.28.1
psycopg2-binary==2.9.2
tqdm==4.64.0
scikit-learn==1.1.2
statsmodels==0.13.2
apscheduler==3.6.3
matplotlib==3.5.1
plotly==5.5.0
ccxt
lxml

pandas_ta
dataframe-image
plotly-resampler
python-telegram-bot
tvdatafeed
flask_restful


